Split agent dashboard into focused modules
